backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app", from = "." },
    { include = "evaluate", from = "." }
]

[tool.poetry.dependencies]
python = "<3.13,>=3.10.0"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
langserve = { extras = ["server"], version = ">=0.2.3" }
numexpr = "^2.10.1"
arxiv = "^2.1.3"
wikipedia = "^1.4.0"
a2wsgi = "^1.10.7"
python-jose = "^3.3.0"
sqlalchemy = "^2.0.35"
psycopg = {version = "^3.2.2", extras = ["binary"]}
python-multipart = "^0.0.9"
langfuse = "^2.50.3"
pubchempy = "^1.0.4"
more-itertools = "^10.5.0"
pydantic-yaml = "^1.3.0"
tabulate = "^0.9.0"
langchain = "^0.2.11"
langchain-community = "^0.2.10"
langchain-milvus = "^0.1.3"
langchain-unstructured = "^0.1.1"
openai = "^1.47.1"
langchain-openai = "^0.1.20"
huggingface-hub = "^0.24.5"
uuid = "^1.30"
chembl-webresource-client = "^0.10.9"

[tool.poetry.group.dev.dependencies]
langchain-cli = "^0.0.21"
black = "^24.8.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
